removed old info menu and dialog files and renamed the new entity files
