+++ demo/demo.c
@@ -10,9 +10,9 @@
 #include "demo/entities/transition.h"
 #include "demo/entities/pause_menu.h"
 #include "demo/entities/fish_menu.h"
-#include "demo/entities/info_menu_entity.h"
+#include "demo/entities/info_menu.h"
 #include "demo/entities/demo_dialog.h"
-#include "demo/entities/info_dialog_entity.h"
+#include "demo/entities/info_dialog.h"
 #include "demo/collision/collision.h"
 #include "demo/scenes/scenes.h"
 #include "demo/menu/menu.h"
