+++ demo/dialog/dialog.h
@@ -22,9 +22,6 @@ void demo_render_dialog(eg_app *, eg_dialog *);
 // new common dialog renderer
 void demo_common_dialog_renderer(eg_app *, eg_entity *);
 
-void demo_init_info_dialog(eg_app *);
-void demo_open_info_dialog(eg_app *);
-
 void demo_init_sign_dialog(eg_app *);
 void demo_open_sign_dialog(eg_app *);
