+++ demo/entities/info_dialog.h
