+++ demo/menu/menu.h
@@ -5,15 +5,6 @@
 
 int demo_init_menus(eg_app *);
 
-void demo_init_pause_menu(eg_app *);
-void demo_open_pause_menu(eg_app *);
-
-void demo_init_fish_menu(eg_app *);
-void demo_open_fish_menu(eg_app *);
-
-void demo_init_info_menu(eg_app *);
-void demo_open_info_menu(eg_app *);
-
 int demo_init_debug_menu(eg_app *);
 void demo_open_debug_menu(eg_app *);
