+++ demo/scenes/scene3.c
@@ -6,8 +6,8 @@
 #include "demo/entities/transition.h"
 #include "demo/entities/pause_menu.h"
 #include "demo/entities/fish_menu.h"
-#include "demo/entities/info_menu_entity.h"
-#include "demo/entities/info_dialog_entity.h"
+#include "demo/entities/info_menu.h"
+#include "demo/entities/info_dialog.h"
 #include "demo/entities/demo_dialog.h"
 
 #include <stdlib.h>
